pokemonManager/requirements.txt
asgiref==3.5.0
attrs==22.1.0
autopep8==1.6.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
coverage==6.5.0
cryptography==37.0.4
defusedxml==0.7.1
Django==3.2.16
django-cors-headers==3.11.0
django-debug-toolbar==3.5.0
django-environ==0.9.0
django-mysql==4.7.0
django-silk==5.0.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
exceptiongroup==1.0.0
flake8==5.0.4
gprof2dot==2022.7.29
idna==3.3
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
mysqlclient==2.1.0
oauthlib==3.2.0
packaging==21.3
pluggy==1.0.0
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.2.0
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-mock==3.10.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.11
